Updating topic behavior
Topics no longer use a straight topic string match and now use
regular expressions.  Currently this code is untested and the
tests are obviously out of date.

Also a bunch of documentation was added to make the design goals
more clear, and to reflect the direction that topic routers are taking.