Registry stub added, and Room routers as well
The stub for the actor registry has been added, this will definitely
change to use a new abstract interface for Actor / Sender etc, but
the goal is to get this working as soon as possible so the remaining
use cases can be worked out.

Room routers are now also mostly functional but untested and well
documented in terms of the design goals.  Rooms provide inexpensive
fanouts to groups of Actors.  At some point, it will need to undergo
a conversion like the Registry (once finished), Topic routers, and
the like.  Actors may end up becoming interface types entirely to
accomodate the collection of things that respond to messages and send.